requirements.txt
beautifulsoup4
jinja2
mistune==0.6
pillow
python-frontmatter==0.2.1
python-slugify
pygments
